requirements.txt
bcrypt==4.1.2
email-validator==2.1.0.post1
fastapi==0.109.0
flake8==7.0.0
psycopg2-binary==2.9.9
pydantic==2.5.3
PyJWT==2.8.0
SQLAlchemy==2.0.25
uvicorn==0.25.0
yagmail==0.15.293
